+++ .claude-plugin/marketplace.json
@@ -14,7 +14,7 @@
         "url": "https://github.com/anam-org/metaxy.git"
       },
       "description": "Feature metadata management for multimodal data and ML pipelines. Provides guidance on feature definitions, versioning, metadata stores, and testing.",
-      "version": "0.1.10",
+      "version": "0.1.11",
       "homepage": "https://anam-org.github.io/metaxy/"
     }
   ]
