+++ src/metaxy/ext/metadata_stores/postgresql.py
@@ -23,7 +23,6 @@ from narwhals.typing import Frame
 from pydantic import Field
 
 from metaxy._decorators import experimental, public
-from metaxy._utils import collect_to_polars
 from metaxy.metadata_store.exceptions import TableNotFoundError
 from metaxy.metadata_store.ibis import IbisMetadataStore, IbisMetadataStoreConfig
 from metaxy.models.constants import (
@@ -32,6 +31,7 @@ from metaxy.models.constants import (
 )
 from metaxy.models.plan import FeaturePlan
 from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.utils import collect_to_polars
 from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
