+++ src/metaxy/metadata_store/ibis.py
@@ -696,7 +696,7 @@ class IbisMetadataStore(MetadataStore, ABC):
 
     def _handle_polars_map_columns(self, df: Frame, map_columns: set[str]) -> Frame:
         """Convert Polars Struct columns to native Arrow Map columns."""
-        from metaxy._utils import collect_to_polars
+        from metaxy.utils import collect_to_polars
         from metaxy.versioning._arrow_map import convert_extension_maps_to_native, convert_structs_to_maps
 
         polars_df = collect_to_polars(df)
