+++ src/metaxy/utils/__init__.py
@@ -1,6 +1,6 @@
 """Utility modules for Metaxy."""
 
-from metaxy._utils import collect_to_polars
 from metaxy.utils.batched_writer import BufferedMetadataWriter
+from metaxy.utils.dataframes import collect_to_arrow, collect_to_polars
 
-__all__ = ["BufferedMetadataWriter", "collect_to_polars"]
+__all__ = ["BufferedMetadataWriter", "collect_to_arrow", "collect_to_polars"]
