+++ tests/ext/lancedb/test_native.py
@@ -9,12 +9,12 @@ import polars as pl
 import pytest
 
 from metaxy import FeatureKey, HashAlgorithm
-from metaxy._utils import collect_to_polars
 from metaxy.config import MetaxyConfig
 from metaxy.ext.metadata_stores.lancedb import LanceDBMetadataStore
 from metaxy.metadata_store import MetadataStore, StoreNotOpenError
 from metaxy.models.feature import FeatureGraph
 from metaxy.models.feature_definition import FeatureDefinition
+from metaxy.utils import collect_to_polars
 from tests.metadata_stores.shared import (
     CRUDTests,
     DeletionTests,
