+++ tests/metadata_stores/core/test_metadata_store.py
@@ -18,13 +18,13 @@ from metaxy import (
     FieldKey,
     FieldSpec,
 )
-from metaxy._utils import collect_to_polars
 from metaxy.ext.metadata_stores.delta import DeltaMetadataStore
 from metaxy.metadata_store import (
     FeatureNotFoundError,
     MetadataSchemaError,
     StoreNotOpenError,
 )
+from metaxy.utils import collect_to_polars
 
 
 @pytest.fixture
