+++ tests/metadata_stores/shared/ibis_map.py
@@ -12,10 +12,10 @@ import ibis
 import pyarrow as pa
 import pytest
 
-from metaxy._utils import collect_to_arrow
 from metaxy.config import MetaxyConfig
 from metaxy.metadata_store import MetadataStore
 from metaxy.models.feature_definition import FeatureDefinition
+from metaxy.utils import collect_to_arrow
 
 
 class IbisMapTests:
