+++ tests/metadata_stores/shared/write.py
@@ -11,7 +11,6 @@ import pytest
 from metaxy_testing.models import SampleFeatureSpec
 
 from metaxy import BaseFeature, FeatureDep, FeatureGraph, FeatureKey, FeatureSpec, FieldDep, FieldKey, FieldSpec
-from metaxy._utils import collect_to_polars
 from metaxy.metadata_store import MetadataStore
 from metaxy.metadata_store.warnings import MetaxyColumnMissingWarning
 from metaxy.models.constants import (
@@ -26,6 +25,7 @@ from metaxy.models.constants import (
     METAXY_PROVENANCE_BY_FIELD,
     METAXY_UPDATED_AT,
 )
+from metaxy.utils import collect_to_polars
 
 
 class WriteTests:
